moved assign disrict to scripts folder
